Moved OCR from Form1 to Ocr Directory
Kinda rough, but first attemt to move everything.
It should be working as before
